refactored code as per requested changes
