+++ controllers/pullRequestsController.js
@@ -1,26 +1,37 @@
 const logger = require('../utils/logger')
-const config = require('config')
 const githubService = require('../services/githubService')
-const { fetchUser } = require('../models/users')
 
 /**
- * Fetches the pull requests in Real-Dev-Squad by user
+ * Loops over an array of objects, takes a value corresponding to key provided and saves it in an array
+ *
+ * @param arrayOfObjects {Array} - Array of objects to loop over
+ * @param key {String} - Value corresponding to this key is saved
+ */
+
+const getNames = (arrayOfObjects, key) => {
+  const names = []
+  arrayOfObjects.forEach((object) => {
+    names.push(object[key])
+  })
+  return names
+}
+
+/**
+ * Collects all pull requests and sends only required data for each pull request
  *
  * @param req {Object} - Express request object
  * @param res {Object} - Express response object
  */
 
-const getPullRequests = async (req, res) => {
+const getPRdetails = async (req, res) => {
   try {
-    const { user } = await fetchUser(req.params.id)
-    const url = `${config.githubApi.baseUrl}/search/issues?q=org:${config.githubApi.org}+author:${user.github_id}+type:pr`
-    const { data } = await githubService.fetch(url)
+    const data = await githubService.fetchPRsByUser(req.params.id)
 
     if (data.total_count) {
       const allPRs = []
       data.items.forEach(({ title, html_url: htmlUrl, state, created_at: createdAt, updated_at: updatedAt, draft, labels, assignees }) => {
-        const allAssignees = githubService.getNames(assignees, 'login')
-        const allLabels = githubService.getNames(labels, 'name')
+        const allAssignees = getNames(assignees, 'login')
+        const allLabels = getNames(labels, 'name')
         allPRs.push({
           title: title,
           url: htmlUrl,
@@ -34,13 +45,14 @@ const getPullRequests = async (req, res) => {
       })
       return res.json(allPRs)
     }
-    return res.json('No pull requests found!')
+    return res.json([])
   } catch (err) {
-    logger.error(`Error while fetching pull requests: ${err}`)
+    logger.error(`Error while processing pull requests: ${err}`)
     return res.boom.badImplementation('Something went wrong please contact admin')
   }
 }
 
 module.exports = {
-  getPullRequests
+  getPRdetails,
+  getNames
 }
